+++ app/requests/Comment_request.java
@@ -7,8 +7,8 @@ public class Comment_request {
 	public Long task_id;
 	public Long comment_id;
 	public String content;
-	public float xpos;
-	public float ypos; 
+	public int xpos;
+	public int ypos; 
 	public int wxpos;
 	public int wypos;
 }
